enforce chineseStyle gameplay for reduced tile sets
There are tile sets, that have only one tile for e.g. the flowers group.
If these tile sets are played in none-chineseStyle, this one tile face
appears too often and not every tile matches another one withe the same
face because they are technically different.
The solution is to enforce chineseStyle gameplay for tile sets that are
known to be reduced.

svn path=/trunk/KDE/kdegames/kshisen/; revision=993597