fix logical error
Checking this option ensures that the created games are solvable.
The option label was wrong. I could have kept the label and
adjusted the code accordingly but I decided to keep the code and
switch the meaning of the label, since it makes more sense
this way.

backport of r1181702, r1181704, r1181709, r1181811 and r1181819

CCBUG: 145953

svn path=/branches/KDE/4.5/kdegames/kshisen/; revision=1182109